requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
greenlet==2.0.1
idna==3.4
numpy==1.24.1
pandas==1.5.2
pyodbc==4.0.35
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
six==1.16.0
spring-config-client==0.2
SQLAlchemy==1.4.46
urllib3==1.26.13
XlsxWriter==3.0.6
